requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.3.0
automated-interpretability==0.0.8
babe==0.0.7
beartype==0.14.1
better-abc==0.0.3
blobfile==2.1.1
boostedblob==0.15.6
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
config2py==0.1.36
contourpy==1.3.1
cycler==0.12.1
datasets==2.21.0
dill==0.3.8
diskcache==5.6.3
docker-pycreds==0.4.0
docstring_parser==0.16
dol==0.3.12
einops==0.8.1
exceptiongroup==1.2.2
fancy-einsum==0.0.3
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
geom_median==0.1.0
gitdb==4.0.12
GitPython==3.1.44
gprof2dot==2024.6.6
graze==0.1.29
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.30.2
i2==0.1.46
idna==3.10
importlib_resources==6.5.2
iniconfig==2.0.0
jaxtyping==0.2.38
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
orjson==3.10.15
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
plotly-express==0.4.1
pluggy==1.5.0
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
py2store==0.1.20
pyarrow==19.0.1
pycryptodomex==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.5
pytest-profiling==1.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.0.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
sae-lens==5.6.0
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sparse_autoencoder==1.10.0
statsmodels==0.14.4
StrEnum==0.4.15
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformer-lens==2.15.0
transformers==4.51.3
transformers-stream-generator==0.0.5
triton==3.2.0
typeguard==4.4.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvloop==0.21.0
wadler_lindig==0.1.4
wandb==0.19.8
xxhash==3.5.0
yarl==1.18.3
zstandard==0.22.0
